2.09.17
forms fixes

+++ CMS.WebUI/Controllers/HomeController.cs
@@ -302,10 +302,10 @@ namespace CMS.WebUI.Controllers
                 {
                     ModelState.AddModelError(key, "Please enter a value for " + key);
                 }
-                if (key != "recaptcha_challenge_field" && key != "recaptcha_response_field")
+                if (key != "recaptcha_challenge_field" && key != "recaptcha_response_field" && key != "x" && key != "y")
                 {
                     formData += FormRepository.RemoveLineEndings(key) + "::" + FormRepository.RemoveLineEndings(Request.Form[key]) + "^^";
-                    emailBody += "<tr><td style='border: 1px solid black; padding-left: 10px; padding-right: 20px;'>" + key + "</td><td style='border: 1px solid black; padding-left: 10px;'>" + Request.Form[key] + "</td></tr>";
+                    emailBody += "<tr style='width: 100%;'><td style='border: 1px solid black; padding-left: 10px; padding-right: 20px; width: 90%'>" + key + "</td><td style='border: 1px solid black; padding-left: 10px;'>" + Request.Form[key] + "</td></tr>";
                 }
             }
